feat(UJS) show helpful error messages for missing components
The Problem
-----------

Currently, if you use an invalid component name in `data-react-class`,
you get a React invariant error that only tells you that the component
class was undefined, but it doesn't actually help you fix the error. I
end up needing to scroll backwards in the stack trace until I find this
part of the `react_ujs_mount.js` code and inspect the value of
`className`

The Solution
------------

Throw a helpful error (and nice console log message) about an invalid
className at the moment we try to mount the component.